add head to front and fix poid, zone management, add contact form, fetch product stock only availble

+++ database/migrations/2024_02_19_022156_create_poids_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('poids', function (Blueprint $table) {
             $table->id();
-            $table->integer('min');
-            $table->integer('max');
+            $table->float('min');
+            $table->float('max');
             $table->timestamps();
         });
     }
